pyproject.toml
[project]
name = "zangetsu-logger"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.37.9",
    "google-cloud-storage>=3.1.0",
    "python-json-logger>=3.2.1",
    "pytz>=2025.1",
    "pyyaml>=6.0.2",
]

# yamlファイルをパッケージに含める
[tool.setuptools.package-data]
"zangetsu_logger" = ["*.yaml"]

[dependency-groups]
dev = [
    "ruff>=0.9.10",
]
